fix: remove import cycle
