+++ src/ai/backend/manager/actions/types.py
@@ -1,0 +1,8 @@
+import enum
+
+
+class OperationStatus(enum.StrEnum):
+    SUCCESS = "success"
+    ERROR = "error"
+    UNKNOWN = "unknown"
+    RUNNING = "running"
